Ensure OAuth2 works
First, this patch updates to zend-expressive-authentication 0.4 and
zend-expressive-authentication-oauth2 0.4.3, as they are the most
recent, stable versions of those libraries; the latter has required
fixes that ensure that the component plays nicely inside an application.

Next, I have updated the ConfigProvider to point to the correct
locations of the OAuth2 artifacts, using the value of `getcwd()` instead
of a path relative to `__DIR__`.

Finally, I have updated the README in a variety of ways:

- I document how to create and populate the sample database and OAuth2
  database. It took me some time to figure that information out!
- I update the documentation for each API call to use proper syntax
  highlighting, to use the `http <METHOD> :8080` syntax (which resolves to
  localhost:8080), and to use all required parameters, including OAuth2
  bearer tokens.
- I add documentation demonstrating sample error conditions.