Provide API documentation for established error types
Problem Details provides a "type" field, which is intended to be a URI
to a human-readable document detailing the type of error.

This patch updates all exception types to use a relative path for the
problem details type URI, and then creates handlers for each that
present text/plain responses with the details.